Move file into adapters folder.
