+++ src/app/auth/operations.js
@@ -1,5 +1,5 @@
 import { hash as namehash } from 'eth-ens-namehash';
-import { rns as registryAddress } from '../configAdapter';
+import { rns as registryAddress } from '../adapters/configAdapter';
 import { checkResolver } from '../notifications';
 
 import {
