+++ src/app/components/ResolverDatalist.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { multilanguage } from 'redux-multilanguage';
-import { publicResolver, multiChainResolver } from '../configAdapter';
+import { publicResolver, multiChainResolver } from '../adapters/configAdapter';
 
 export default multilanguage(({ strings }) => (
   <datalist id="resolvers">
