+++ src/app/tabs/AdminMyCryptoTab/rns.js
@@ -2,7 +2,7 @@ import Web3 from 'web3';
 import { hash as namehash } from 'eth-ens-namehash';
 import { rskMain } from '../../../config/nodes.json';
 import { rnsAbi } from './abis.json';
-import { rns as rnsAddress } from '../../configAdapter';
+import { rns as rnsAddress } from '../../adapters/configAdapter';
 
 export const getRnsField = (field, name) => {
   const web3 = new Web3(rskMain);
